Partial error handling added

+++ src/main/java/com/JAZ1/zad1/PotionsService.java
@@ -1,5 +1,8 @@
 package com.JAZ1.zad1;
 
+import com.JAZ1.zad1.exeptions.BadPotionFieldsException;
+import com.JAZ1.zad1.exeptions.BadPotionPriceException;
+import com.JAZ1.zad1.exeptions.PotionNotFoundException;
 import com.JAZ1.zad1.model.Potion;
 import com.baeldung.openapi.model.PotionsCreateRequest;
 import com.baeldung.openapi.model.PotionsResponse;
@@ -18,6 +21,8 @@ public class PotionsService {
     private final PotionsMapper mapper;
 
     public PotionsResponse addPotion(PotionsCreateRequest request) {
+
+        verifyPotionFields(request);
         Potion saved = repository.save(mapper.mapToPotion(request));
 
         return mapper.mapToPotionsResponse(saved);
@@ -28,11 +33,15 @@ public class PotionsService {
     }
 
     public PotionsResponse getPotion(UUID id) {
-        Potion saved = repository.getReferenceById(id);
-        return mapper.mapToPotionsResponse(saved);
+        if (!repository.existsById(id))
+            throw new PotionNotFoundException("no potion under the id: " + id);
+        return mapper.mapToPotionsResponse(repository.getReferenceById(id));
     }
 
     public PotionsResponse updatePotion(UUID id, PotionsCreateRequest req) {
+        if (!repository.existsById(id))
+            throw new PotionNotFoundException("no potion under the id: " + id);
+        verifyPotionFields(req);
         Potion potion = repository.getReferenceById(id);
         potion.setName(req.getName());
         potion.setPrice(req.getPrice());
@@ -43,7 +52,18 @@ public class PotionsService {
     }
 
     public void deletePotion(UUID id) {
+        if (!repository.existsById(id))
+            throw new PotionNotFoundException("no potion under the id: " + id);
         repository.deleteById(id);
+    }
+
+    private void verifyPotionFields(PotionsCreateRequest req) {
+        if (req.getPrice() < 0.0)
+            throw new BadPotionPriceException();
+        if (req.getDescription().isEmpty())
+            throw new BadPotionFieldsException();
+        if (req.getName().isEmpty())
+            throw new BadPotionFieldsException();
     }
